requirements.txt
pydantic>=2.5.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
faker>=20.0.0
typing-extensions>=4.8.0
rich>=13.7.0 